Order

+++ src/main/java/cafe_Management/demo/Repository/OrderRepository.java
@@ -1,9 +1,12 @@
 package cafe_Management.demo.Repository;
 
-import jakarta.persistence.criteria.Order;
+import cafe_Management.demo.enitites.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
+
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
+
 }
